Moved tool bar to separate component
